[jsk_recognition_utils] Separate convex_polygon.h from geo_util.h
